Hiking Mobile App
Overview 
As a co-founder of a startup company, you want to develop a mobile application with your team  members. For most industries a mobile strategy is no longer an option, it is a requirement. In the  world of mobile app development, it is vital to success to get it right the first time. One of the top  reasons for mobile apps failing at the market within the first few months is due to poor planning  
and lack of strategy. The mobile app development process begins with the concept. From there,  an app can go through many different stages of evolution. As such, even a small perspective  diversion during the process can dramatically change the development and end product. So,  system analysis and design (SAD) of app development helps alleviate this risk. 
When developing an app, the SAD can often be rushed and even at times, completely overlooked.  The more questions you can answer about the app and its functionality at the start, the stronger  the development process will be. Spending adequate time and effort on pre-planning allows you  to not only define the core features of your app but also plan for unexpected turning points in  development. Successful SAD can also help identify the goals for the app, early insights, and  allow you to prepare an effective competitor and marketing strategy.